Split up gammaray_probe
Use two libraries approach:
* Gammaray Core: core library
* Gammaray Probe: probe library (with hooks) linked to core library

Benefits:
* Plugins may link to gammaray_core directly
* We can write unit tests that link against gammaray_core
  Benefit: No hooking mechanisms triggered then
* Clearer separation of what belongs to hooking

It is mostly moving around of code, not too many semantic changes.
There *might* be issues wrt multithreading / missing locks.

New class: SignalSlotsLocationStore for storing and retrieving a
connect() location.
